Refactored osp add nodes to use a different modal.
