+++ fusor-ember-cli/app/routes/deployment.js
@@ -1,6 +1,6 @@
 import Ember from 'ember';
 import DeploymentRouteMixin from "../mixins/deployment-route-mixin";
-import OspNodeManager from "../models/osp-node-manager";
+import OspNodeManager from "../utils/osp/osp-node-manager";
 import request from 'ic-ajax';
 
 export default Ember.Route.extend(DeploymentRouteMixin, {
