+++ fusor-ember-cli/app/styles/openstack.scss
@@ -517,6 +517,13 @@
   margin-bottom: 10px;
 }
 
+.add-node-form {
+  margin-top: 16px;
+}
+
+.add-node-uneditable-input-container {
+  padding-top: 2px;
+}
 .new-node-registration-csv-errors {
   color: $error-color;
 }
