+++ fusor-ember-cli/app/templates/register-nodes.hbs
@@ -101,4 +101,5 @@
                     deploymentName=deploymentName}}
 
 {{partial 'new-node-registration'}}
+{{partial 'add-node-registration'}}
 {{partial 'delete-node-confirmation'}}
