+++ fusor-ember-cli/app/utils/osp/osp-node-manager.js
